+++ src/main/java/henry/TaskList.java
@@ -1,3 +1,6 @@
+package henry;
+
+import henry.task.Task;
 import java.util.ArrayList;
 
 public class TaskList {
