+++ src/main/java/henry/Ui.java
@@ -1,3 +1,5 @@
+package henry;
+
 import java.util.Scanner;
 
 public class Ui {
