+++ src/main/java/henry/command/ByeCommand.java
@@ -1,3 +1,10 @@
+package henry.command;
+
+import henry.Storage;
+import henry.TaskList;
+import henry.Ui;
+import henry.HenryException;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws HenryException {
