+++ src/main/java/henry/task/Deadline.java
@@ -1,6 +1,9 @@
+package henry.task;
+
+import henry.HenryException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-class Deadline extends Task {
+public class Deadline extends Task {
     private final LocalDateTime date;
     private static final DateTimeFormatter INPUT_FORMATTER = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
     private static final DateTimeFormatter OUTPUT_FORMATTER = DateTimeFormatter.ofPattern("MMM dd yyyy HH:mm");
